refactor bsp spec into separate files
[ci skip-rust]

[ci skip-build-wheels]
