+++ src/python/pants/bsp/context.py
@@ -6,7 +6,8 @@ import threading
 from pathlib import Path
 from typing import Callable
 
-from pants.bsp.spec import BSPNotification, InitializeBuildParams
+from pants.bsp.spec.lifecycle import InitializeBuildParams
+from pants.bsp.spec.task import BSPNotification
 
 # Wrapper type to provide BSP rules with the ability to interact with the BSP protocol driver.
 #
