+++ src/python/pants/jvm/bsp/spec.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Any
 
-from pants.bsp.spec import Uri
+from pants.bsp.spec.base import Uri
 
 
 @dataclass(frozen=True)
